requirements.txt
aiosmtplib==2.0.2
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.3
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-mail==1.4.1
fonttools==4.43.1
geographiclib==2.0
geopy==2.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.3
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
passlib==1.7.4
Pillow==10.0.1
psycopg2-binary==2.9.8
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
XlsxWriter==3.1.9
xlwings==0.30.12
zipp==3.17.0
